[BEAM-9748] Refactor Reparallelize as an alternative Reshuffle implementation
